fix: bug fix calculator operations functions

+++ Arman/Lesson_9/calculator/calculator.py
@@ -17,14 +17,13 @@ print("3. Multiplication  (*)")
 print("4. Division        (/)")
 
 # ask for first calculator number in1 console from user
-ask_for_first_number = int(input("Enter first number: "))
+ask_for_first_number = float(input("Enter first number: "))
 # ask for first calculator number in console from user
-ask_for_second_number = int(input("Enter second number: "))
+ask_for_second_number = float(input("Enter second number: "))
 # ask for first calculator number in console from user
 ask_for_operation_type = input("Enter operation: ")
 # run calculator operation function and print result to console
-calculator_operation(int(ask_for_first_number), int(ask_for_second_number), ask_for_operation_type)
-
+calculator_operation(ask_for_first_number, ask_for_second_number, ask_for_operation_type)
 
 # Anna - interesting and complex solution
 # Your calculator let input only integers, in case I try input float, get error, 
@@ -33,6 +32,6 @@ calculator_operation(int(ask_for_first_number), int(ask_for_second_number), ask_
 # int(10.1) -> 10
 # int("10.1")-> Invalid value error
 
-# And In case we divide on 0, you cold return error message, not 0
+# And In case we divide on 0, you could return error message, not 0
 
 # The code style very good - code is structured and readable!!!
